requirements.txt
opencv-python==4.10.0.84
numpy==2.1.3
scikit-image==0.24.0
scipy==1.14.1
Pillow==11.0.0



